+++ Kernel/System/Stats/Dynamic/TicketSolutionResponseTime.pm
@@ -39,8 +39,6 @@ sub new {
     my $Self = {};
     bless( $Self, $Type );
 
-    $Self->{DBSlaveObject} = $Param{DBSlaveObject} || $Kernel::OM->Get('Kernel::System::DB');
-
     # get the dynamic fields for ticket object
     $Self->{DynamicField} = $Kernel::OM->Get('Kernel::System::DynamicField')->DynamicFieldListGet(
         Valid      => 1,
@@ -485,17 +483,19 @@ sub GetObjectAttributes {
         push @ObjectAttributes, @ObjectAttributeAdd;
     }
 
+    my $DBObject = $Kernel::OM->Get('Kernel::System::DB');
+
     if ( $ConfigObject->Get('Stats::CustomerIDAsMultiSelect') ) {
 
         # Get CustomerID
         # (This way also can be the solution for the CustomerUserID)
-        $Self->{DBSlaveObject}->Prepare(
+        $DBObject->Prepare(
             SQL => "SELECT DISTINCT customer_id FROM ticket",
         );
 
         # fetch the result
         my %CustomerID;
-        while ( my @Row = $Self->{DBSlaveObject}->FetchrowArray() ) {
+        while ( my @Row = $DBObject->FetchrowArray() ) {
             if ( $Row[0] ) {
                 $CustomerID{ $Row[0] } = $Row[0];
             }
@@ -912,6 +912,8 @@ sub _ReportingValues {
     my ( $Self, %Param ) = @_;
     my $SearchAttributes = $Param{SearchAttributes};
 
+    my $DBObject = $Kernel::OM->Get('Kernel::System::DB');
+
     #
     # escape search attributes for ticket search
     #
@@ -947,13 +949,13 @@ sub _ReportingValues {
         if ( ref $TicketSearch{$Attribute} ) {
             if ( ref $TicketSearch{$Attribute} eq 'ARRAY' ) {
                 $TicketSearch{$Attribute} = [
-                    map { $Self->{DBSlaveObject}->QueryStringEscape( QueryString => $_ ) }
+                    map { $DBObject->QueryStringEscape( QueryString => $_ ) }
                         @{ $TicketSearch{$Attribute} }
                 ];
             }
         }
         else {
-            $TicketSearch{$Attribute} = $Self->{DBSlaveObject}->QueryStringEscape(
+            $TicketSearch{$Attribute} = $DBObject->QueryStringEscape(
                 QueryString => $TicketSearch{$Attribute}
             );
         }